sample-plugin/test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

hacking<0.11,>=0.10.0

#yank this for now. We aren't doing unit testing. I'm sure we
#are going to have problem with stable/liberty and master coexistance
#at some point!
#http://tarballs.openstack.org/horizon/horizon-master.tar.gz#egg=horizon
coverage>=3.6 # Apache-2.0
ddt>=1.0.1 # MIT
discover # BSD
django-nose>=1.2 # BSD
python-subunit>=0.0.18 # Apache-2.0/BSD
nose-exclude # LGPL
selenium>=2.50.1 # Apache-2.0
sphinx!=1.2.0,!=1.3b1,<1.3,>=1.1.2 # BSD
oslosphinx!=3.4.0,>=2.5.0 # Apache-2.0
oslotest>=1.10.0 # Apache-2.0
testrepository>=0.0.18 # Apache-2.0/BSD
testscenarios>=0.4 # Apache-2.0/BSD
testtools>=1.4.0 # MIT
# This also needs xvfb library installed on your OS
xvfbwrapper>=0.1.3 #license: MIT
